refactor(spec-546): name the modules after their contents — four modules out of one drawer (#674)
agent/handlers/shared.ts — one 1531-line file named after the act that created
it — becomes four modules named after their contents.

  tool-contract.ts     436 lines, 12 exports — the shape every tool declares
                       and the context every handler receives
  doc-state.ts         102 lines,  3 exports — load a doc's state, render it
  guidance-envelope.ts 927 lines,  8 exports — one public entry point
  related-issues.ts    111 lines,  3 exports — semantic issue overlap search

Selection rule [std-51 cl-9]: the name must be true of EVERY symbol in the
module, applied per module rather than by picking a module count. That is what
produced four rather than three — related-issues does semantic search, and
folding it into guidance-envelope would make that name false of five of its
symbols.

`git mv` for the rename, so `git log --follow tool-contract.ts` reaches
08bf9ce0 "spec-366: extract tool-specs.ts handler infra into handlers/shared.ts".

26 exports across the four, up from 24: FooterSignal and FullDocState are
promoted because they now cross a module boundary and so have a consumer
outside their file — precisely when std-51 permits an export.

A cycle removed rather than papered over. FooterSignal declared
`issueHits: Awaited<ReturnType<typeof relatedIssuesForDecision>>`, making
tool-contract depend on related-issues while related-issues depended on
ToolCtx/ResolvedRef there. An `import type` would have compiled (types are
erased, so no runtime cycle) but left a cycle in the graph for every reader.
That derived type resolves to exactly MemexSearchHit[], so naming it directly
deletes the edge. Types only; it also fixed 8 implicit-any errors in two test
files that came from inference through the indirection.

Three guard edits, deliberate; two were forced and would otherwise throw
ENOENT at module load rather than fail an assertion:

- guidance-authoring-confined and spec-249-status-overview each read
  shared.ts by name FIRST, then excluded it from the readdir, purely so the
  confinement window's two anchors appeared in order. Both anchors now live in
  one file, which preserves their order itself, so the hack is gone rather
  than repointed. Added .sort() while there: readdirSync order is
  filesystem-dependent and both guards compute a window from positions.
- response-budget-funnel's hardcoded caller list now reads
  ["agent/handlers/doc-state.ts"] — the list finally states a fact instead of
  pointing at a drawer.

No assertion deleted, none relaxed, no file skipped.

Two corrections to the Spec's own plan, recorded in dec-1 rather than hidden:
constraint (c) required the two window anchors to be "adjacent", which is
wrong — composeSupersessionHeader sits between them and the guard works,
because the window is simply the text BETWEEN the anchors; the real constraint
is ORDER. And FooterSignal/FooterSlot went to tool-contract, not
guidance-envelope, since ToolCtx.footerSlot makes them the shape of a ctx field.

No test added, deliberately: a test asserting a module list or export count
would gate interface shape, which std-51 cl-34 forbids and this Spec lists as
out of scope.

Verified on the rebased tree, not carried over: make typecheck exit 0; real
.d.ts emission exit 0 separately, since --noEmit skips declaration emit; zero
unused import or local in the four modules; the confinement window measured at
9314 characters, BYTE-IDENTICAL in size to before the split, with all three
functions inside it in order; full server suite 707 files / 6558 tests passed;
the seven source-scanning guards run explicitly, 45 tests passed.

The suite's +6 cases were enumerated with `vitest list` and diffed rather than
guessed at: mutate-coverage.static-scan (std-8) and no-direct-anthropic
(std-30) generate one case per source file, so each new module picks up both.
Nothing was lost.

Spec: https://memex.ai/mindset-prod/memex-building-itself/specs/spec-546
