+++ packages/server/src/agent/handlers/acs.ts
@@ -36,13 +36,15 @@ import {
   ValidationError,
 } from "../../types/errors.js";
 import {
-  VERBOSE_FIELD,
   formatAcCoverageSummary,
+} from "./guidance-envelope.js";
+import {
+  VERBOSE_FIELD,
   isDocLikeKind,
   reqCtx,
   resolveRefArg,
   type ToolSpec,
-} from "./shared.js";
+} from "./tool-contract.js";
 
 // ── Moved here from shared.ts by spec-546 t-2: this file is the symbol's only
 // consumer, so it lives with its consumer and is private [per std-51].
