+++ packages/server/src/agent/handlers/tasks.ts
@@ -35,14 +35,16 @@ import {
   formatDocStatusHeader,
 } from "../../formatting/formatters.js";
 import {
-  VERBOSE_FIELD,
   formatState,
   fullDocState,
+} from "./doc-state.js";
+import {
+  VERBOSE_FIELD,
   isDocLikeKind,
   reqCtx,
   resolveRefArg,
   type ToolSpec,
-} from "./shared.js";
+} from "./tool-contract.js";
 import { eq } from "drizzle-orm";
 import { db } from "../../db/connection.js";
 import { taskDeps } from "../../db/schema.js";
