+++ packages/server/src/agent/spec-249-status-overview.integration.test.ts
@@ -168,17 +168,21 @@ describe("ac-7 — no new tool/flag; verbose stays full state; read-path only", 
 });
 
 describe("ac-6 — the overview is composed in the single seat (source guard)", () => {
-  // spec-366: composeGuidanceEnvelope + craftStatusOverview moved to
-  // agent/handlers/shared.ts; the per-tool handlers to agent/handlers/*.ts. Scan
-  // shared.ts (the seat) FIRST, then append the other handlers so any handler
+  // spec-366 / spec-546: composeGuidanceEnvelope + craftStatusOverview live
+  // together in ONE module under agent/handlers/; the per-tool handlers are the
+  // other agent/handlers/*.ts. Concatenate every handler module so any handler
   // referencing craftStatusOverview directly is flagged outside the seat.
+  //
+  // spec-546: this used to read shared.ts by name FIRST so the seat's anchors
+  // appeared in order. Unnecessary now that both live in the same file, which
+  // preserves their order itself. Sorted for determinism — readdirSync order is
+  // filesystem-dependent and the seat is computed from positions.
   const handlersDir = join(__dirname, "handlers");
-  const SRC = [
-    readFileSync(join(handlersDir, "shared.ts"), "utf-8"),
-    ...readdirSync(handlersDir)
-      .filter((n) => n.endsWith(".ts") && n !== "shared.ts")
-      .map((n) => readFileSync(join(handlersDir, n), "utf-8")),
-  ].join("\n");
+  const SRC = readdirSync(handlersDir)
+    .filter((n) => n.endsWith(".ts"))
+    .sort()
+    .map((n) => readFileSync(join(handlersDir, n), "utf-8"))
+    .join("\n");
 
   it("craftStatusOverview is referenced only inside composeGuidanceEnvelope (or its own def)", () => {
     tagAc(AC(6));
